+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adds support for one-call-buys on shipments and orders via adding the `service` key to both objects
 * Removes the unused `orderBy` parameter from the `Batch` object
 * Update the `DefaultApiBase` to include `v2` and remove `v2` from every request url string
+* Add a 30 second connection timeout and a 60 second request timeout for all HTTP requests
 
 ## 2.7.0 (2021-11-24)